* GSWeb.framework/GSWMultiKeyDictionary.h/m: 	Adapt API to current NSDictionary API.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gsweb/trunk@30231 72102866-910b-0410-8b05-ffd578937521

+++ GSWeb.framework/GSWMultiKeyDictionary.h
@@ -40,7 +40,7 @@
 };
 
 +(id)dictionary;
--(id)initWithCapacity:(unsigned int)capacity;
+-(id)initWithCapacity:(NSUInteger)capacity;
 
 -(void)setObject:(id)object
          forKeys:(id)keys,...;
